+++ neo/renderer/Material.h
@@ -3,7 +3,7 @@
 
 Doom 3 BFG Edition GPL Source Code
 Copyright (C) 1993-2012 id Software LLC, a ZeniMax Media company.
-Copyright (C) 2014-2016 Robert Beckebans
+Copyright (C) 2014-2020 Robert Beckebans
 Copyright (C) 2014-2016 Kot in Action Creative Artel
 
 This file is part of the Doom 3 BFG Edition GPL Source Code ("Doom 3 BFG Edition Source Code").
@@ -498,6 +498,14 @@ public:
 		return unsmoothedTangents;
 	}
 
+	// RB: characters and models that baked in Blender or Substance designer use the newer
+	// Mikkelsen tangent space standard.
+	// see: https://bgolus.medium.com/generating-perfect-normal-maps-for-unity-f929e673fc57
+	bool				UseMikkTSpace() const
+	{
+		return mikktspace;
+	}
+
 	// by default, monsters can have blood overlays placed on them, but this can
 	// be overrided on a per-material basis with the "noOverlays" material command.
 	// This will always return false for translucent surfaces
@@ -891,6 +899,7 @@ private:
 	bool				blendLight;
 	bool				ambientLight;
 	bool				unsmoothedTangents;
+	bool				mikktspace;			// RB: use Mikkelsen tangent space standard for normal mapping
 	bool				hasSubview;			// mirror, remote render, etc
 	bool				allowOverlays;