+++ RELEASE-NOTES.md
@@ -95,6 +95,8 @@ The main goal is that the new content looks the same in RBDOOM-3-BFG as in Blend
 
 * Bumped the static vertex cache limit of 31 MB to roughly ~ 128 MB to help with some custom models and maps by the Doom 3 community
 
+* Added support for Mikkelsen tangent space standard for new assets (thanks to Stephen Pridham)
+
 
 
 _______________________________________